# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Bitcoin Wallet API"
  },
  "host": "acube.tech",
  "basePath": "/bitcoinAPI",
  "paths": {
    "/transaction/{nickname}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Get all the transaction from a nickname",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nickname",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Transaction"
              }
            }
          }
        }
      }
    },
    "/transaction": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "summary": "add a new transaction",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "transaction to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Transaction"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Transaction"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Transaction": {
      "type": "object",
      "properties": {
        "owner": {
          "type": "string"
        },
        "timestamp": {
          "type": "string"
        },
        "usd": {
          "type": "number",
          "format": "double"
        },
        "btc": {
          "type": "number",
          "format": "double"
        },
        "type": {
          "type": "boolean"
        }
      }
    }
  }
}